partial transition to structure Feature-Sliced Design, Bugs fixed
