+++ package.json
@@ -21,11 +21,15 @@
     "@mui/material": "^6.1.6",
     "@reduxjs/toolkit": "^2.3.0",
     "@supabase/supabase-js": "^2.46.1",
+    "css-loader": "^7.1.2",
+    "immer": "^10.1.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-redux": "^9.1.2",
     "react-router-dom": "^6.28.0",
     "sass": "^1.80.6",
+    "sass-loader": "^16.0.3",
+    "style-loader": "^4.0.0",
     "vitest": "^2.1.4"
   },
   "devDependencies": {
