+++ src/App.tsx
@@ -1,9 +1,15 @@
-import Index from "./shared/api/auth";
+import { Routes, Route } from "react-router-dom";
+import LoginPage from "./pages/LoginPgae/LoginPage";
+import SignUpPage from "./pages/SignUpPage/SignUpPage";
+
 
 const App = () => {
 
   return (
-    <Index />
+    <Routes>
+    <Route path="/" element={<LoginPage />} />
+    <Route path="/signUp" element={<SignUpPage />} />
+</Routes>
   );
 };
