+++ vite.config.ts
@@ -1,7 +1,9 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
 
-// https://vite.dev/config/
+
+// https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
-})
+
+});
